refactor: ListItem을 list 위치로 이동
