+++ src/pages/list/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Header from "../../components/Header";
-import ListItem from "../../components/ListItem";
+import ListItem from "./ListItem";
 import places from "../../data/places";
 
 export default function ListPage() {
